+++ src/MarkdownMate.ts
@@ -1,4 +1,5 @@
 import MarkdownEditor from "./lib/MarkdownEditor";
+import Table from "./lib/table";
 import Toc from "./lib/toc/index";
 
 export default class MarkdownMate {
@@ -35,5 +36,50 @@ export default class MarkdownMate {
   public toggleBlockquote() {
     const markdownEditor = new MarkdownEditor();
     markdownEditor.toggleBlockquote();
+  }
+
+  public insertTable() {
+    const table = new Table();
+    table.insert();
+  }
+
+  public formatTable() {
+    const table = new Table();
+    table.format();
+  }
+
+  public insertRowBefore() {
+    const table = new Table();
+    table.insertRowBeforeCurrentRow();
+  }
+
+  public insertRowAfter() {
+    const table = new Table();
+    table.insertRowAfterCurrentRow();
+  }
+
+  public insertColumnBefore() {
+    const table = new Table();
+    table.insertColumnBeforeCurrentColumn();
+  }
+
+  public insertColumnAfter() {
+    const table = new Table();
+    table.insertColumnAfterCurrentColumn();
+  }
+
+  public deleteCurrentColumn() {
+    const table = new Table();
+    table.deleteCurrentColumn();
+  }
+
+  public deleteCurrentRow() {
+    const table = new Table();
+    table.deleteCurrentRow();
+  }
+
+  public changeColumnAlignment() {
+    const table = new Table();
+    table.changeColumnAlignment();
   }
 }