+++ src/lib/table/enums.ts
@@ -1,0 +1,5 @@
+export enum TableColumnAlignment {
+  LEFT,
+  CENTER,
+  RIGHT,
+}